Rebuild most function
Fix the bug that can not restore the same master key from the same seed.
And now, the seed will be fix 256bit. The seed will extend via sha3.Sum256 if need.
And most function has be change to method of corresponding struct.
And make a new file named util.go to put the help function.
Modify the error handling in Sign, Verify, Link function.
Delete the Serialize and Deserialize of seed, because it can be substitute with hex.EncodeToString and hex.DecodeString.
At last, we add test to check the correctness.